Revert "better additive increase"
This reverts commit 3cc18a0d32d53abbd215442d2f7e4b28bf524d7f.

+++ project1/sourdough/datagrump/controller.hh
@@ -8,8 +8,8 @@
 class Controller {
 private:
     bool debug_; /* Enables debugging output */
-    double cwnd_;
-    double thresh_;
+    uint64_t cwnd_;
+    uint64_t thresh_;
 
     /* Add member variables here */
